+++ EventPlanner/src/app/app.module.ts
@@ -3,50 +3,73 @@ import { BrowserModule } from '@angular/platform-browser';
 import { RouterModule } from '@angular/router';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations'; 
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { LayoutModule } from './layout/layout.module';
 import { OfferingModule } from './offering/offering.module';
 import { WineModule } from './wine/wine.module';
-import { MatButtonModule } from '@angular/material/button';
 import { EventModule } from './event/event.module';
-import { CreateCategoryDialogComponent } from './offering/create-category-dialog/create-category-dialog.component';
-import { MatIconModule } from '@angular/material/icon'; 
+import { NotificationModule } from './notification/notification.module';
+
+// Material Imports
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
-import { FormsModule } from '@angular/forms';
 import { MatDialogModule } from '@angular/material/dialog';
 import { MatSelectModule } from '@angular/material/select';
 import { MatSliderModule } from '@angular/material/slider';
-import { ReactiveFormsModule } from '@angular/forms';
-import {MatDatepickerModule} from '@angular/material/datepicker';
-import { NotificationModule } from './notification/notification.module';
+import { MatDatepickerModule } from '@angular/material/datepicker';
+import { MatCardModule } from '@angular/material/card';
+import { MatTableModule } from '@angular/material/table';
+import { MatPaginatorModule } from '@angular/material/paginator';
+
+// Forms Imports
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+
+// Components
+import { CreateCategoryDialogComponent } from './offering/create-category-dialog/create-category-dialog.component';
+import { CategoryDialogComponent } from './offering/category-dialog/category-dialog.component';
+import { OfferingCategoryComponent } from './offering/offering-category/offering-category.component';
+
+// Services
+import { CategoryService } from './offering/category-service/category.service';
 
 @NgModule({
   declarations: [
     AppComponent,
     CreateCategoryDialogComponent,
+    CategoryDialogComponent,
+    OfferingCategoryComponent
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
     LayoutModule,
     WineModule,
-    MatButtonModule,
     EventModule,
     OfferingModule,
-    MatIconModule, 
+    NotificationModule,
+
+    // Material Modules
+    MatButtonModule,
+    MatIconModule,
     MatFormFieldModule,
     MatInputModule,
-    FormsModule,
     MatDialogModule,
-    BrowserAnimationsModule,
     MatSelectModule,
     MatSliderModule,
-    ReactiveFormsModule,
     MatDatepickerModule,
-    NotificationModule
+    MatCardModule,
+    MatTableModule,
+    MatPaginatorModule,
+
+    // Animation and Forms
+    BrowserAnimationsModule,
+    FormsModule,
+    ReactiveFormsModule
   ],
   providers: [
+    CategoryService
   ],
   bootstrap: [AppComponent]
 })
