+++ EventPlanner/src/app/layout/home/home.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Event } from '../../event/model/event.model';
 import { EventService } from '../../event/event.service';
-import { OfferingService } from '../../offering/offering.service';
+import { OfferingService } from '../../offering/offering-service/offering.service';
 import { Offering } from '../../offering/model/offering.model';
 import { FilterEventsDialogComponent } from '../../event/filter-events-dialog/filter-events-dialog.component';
 import { MatDialog } from '@angular/material/dialog';
