+++ EventPlanner/src/app/layout/nav-bar/nav-bar.component.html
@@ -10,5 +10,8 @@
       <div class="navigation-item">
         <mat-icon [routerLink]="['notification-panel']">notifications</mat-icon>
       </div>
+      <div class="navigation-item">
+        <mat-icon [routerLink]="['offering-categories']">category</mat-icon>
+      </div>
     </div>
 </mat-toolbar>
