--- EventPlanner/src/app/offering/comment.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { CommentService } from './comment.service';
-
-describe('CommentService', () => {
-  let service: CommentService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(CommentService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
