+++ EventPlanner/src/app/offering/comments/comment.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { CommentService } from '../comment.service';
+import { CommentService } from '../comment-service/comment.service';
 describe('CommentService', () => {
   let service: CommentService;
