+++ EventPlanner/src/app/offering/create-category-dialog/create-category-dialog.component.css
@@ -1,57 +1,64 @@
 mat-dialog-content {
-    background-color: #ebe3e5;
-    padding: 24px;
-    border-radius: 8px;
-    color: #252629;
-  }
-  
-  mat-dialog-content header h1 {
-    font-size: 1.8em;
-    margin: 0 0 16px 0;
-    text-align: center;
-    color: #252629;
-  }
-  
-  mat-dialog-content p {
-    margin: 16px 0;
-  }
-  
-  mat-form-field {
-    width: 100%;
-  }
-  
-  mat-form-field input {
-    color: #252629;
-  }
-  
-  mat-dialog-actions {
-    padding: 16px;
-    text-align: center;
-  }
-  
-  button[mat-raised-button] {
-    margin: 0 8px;
-    padding: 8px 24px;
-    border-radius: 4px;
-  }
-  
-  button[mat-raised-button][color="primary"] {
-    background-color: #f8adbb;
-    color: #252629;
-    transition: background-color 0.3s ease-in-out;
-  }
-  
-  button[mat-raised-button][color="primary"]:hover {
-    background-color: #e097a8;
-  }
-  
-  button[mat-raised-button][color="warn"] {
-    background-color: #252629;
-    color: #ebe3e5;
-    transition: background-color 0.3s ease-in-out;
-  }
-  
-  button[mat-raised-button][color="warn"]:hover {
-    background-color: #1e1e20;
-  }
-  
+  padding: 24px;
+  border-radius: 8px;
+  color: #252629;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+mat-dialog-content header h1 {
+  font-size: 1.8em;
+  margin: 0 0 24px 0;  
+  text-align: center;
+  color: #252629;
+}
+
+.form-container {
+  width: 100%;
+  max-width: 400px;  
+  margin-bottom: 24px;  
+}
+
+mat-form-field {
+  width: 100%;
+}
+
+mat-form-field input {
+  color: #252629;
+}
+
+mat-dialog-actions {
+  padding: 16px;
+  text-align: center;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+}
+
+button[mat-raised-button] {
+  margin: 0 8px;
+  padding: 8px 24px;
+  border-radius: 4px;
+}
+
+button[mat-raised-button][color="primary"] {
+  background-color: #bebce6;
+  color: white;
+  transition: background-color 0.3s ease-in-out;
+}
+
+button[mat-raised-button][color="primary"]:hover {
+  background-color: #a9abc9;
+}
+
+button[mat-raised-button][color="warn"] {
+  background-color: #252629;
+  color: #ebe3e5;
+  transition: background-color 0.3s ease-in-out;
+}
+
+button[mat-raised-button][color="warn"]:hover {
+  background-color: #1e1e20;
+}
