+++ EventPlanner/src/app/offering/create-category-dialog/create-category-dialog.component.html
@@ -6,9 +6,10 @@
         <mat-form-field appearance="fill">
             <mat-label>Category Name</mat-label>
             <input matInput placeholder="Enter category name" />
-        </mat-form-field>          
+        </mat-form-field>
     </p>
 </mat-dialog-content>
+
 <mat-dialog-actions>
     <button mat-raised-button color="primary" matDialogSave>Save</button>
     <button mat-raised-button color="warn" matDialogClose>Cancel</button>
