+++ EventPlanner/src/app/offering/create-offerings/create-offerings.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { MatDialog } from '@angular/material/dialog';
 import { CreateCategoryDialogComponent } from '../create-category-dialog/create-category-dialog.component';
-import { OfferingService } from '../offering.service';
+import { OfferingService } from '../offering-service/offering.service';
 
 @Component({
   selector: 'app-create-offerings',
