+++ EventPlanner/src/app/offering/details-page/details-page.component.ts
@@ -5,8 +5,8 @@ import { FormsModule } from '@angular/forms';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { ActivatedRoute } from '@angular/router';
-import { OfferingService } from '../offering.service';
-import { CommentService, Comment } from '../comment.service';
+import { OfferingService } from '../offering-service/offering.service';
+import { CommentService, Comment } from '../comment-service/comment.service';
 import { Offering } from '../model/offering.model';
 import { switchMap } from 'rxjs/operators';
 import { Service } from '../model/service.model';
