--- EventPlanner/src/app/offering/offering.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { OfferingService } from './offering.service';
-
-describe('OfferingService', () => {
-  let service: OfferingService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(OfferingService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
